LPD-48502 Move to folder
